requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastapi==0.95.1
Flask==2.3.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.13.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
langchain==0.0.153
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.24.3
openai==0.27.5
openapi-schema-pydantic==1.2.4
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
pyzmq==25.0.2
requests==2.29.0
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
sniffio==1.3.0
SQLAlchemy==2.0.12
stack-data==0.6.2
starlette==0.26.1
tenacity==8.2.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
wcwidth==0.2.6
Werkzeug==2.3.2
yarl==1.9.2
zipp==3.15.0
